Collapse observable source implementation
