+++ test/watch/nestedWatchSpec.ts
@@ -2,7 +2,7 @@ import { PassThroughFork } from "../../src/fork/pass-through-fork";
 import { ensure, Jinaga } from "../../src/jinaga";
 import { FactManager } from "../../src/managers/factManager";
 import { MemoryStore } from "../../src/memory/memory-store";
-import { ObservableSourceImpl } from "../../src/observable/observable-source-impl";
+import { ObservableSource } from "../../src/observable/observable";
 import { MockAuthentication } from "./mock-authentication";
 
 interface Room {
@@ -51,7 +51,7 @@ describe("Nested watch", () => {
 
     beforeEach(() => {
         const memory = new MemoryStore();
-        const observableSource = new ObservableSourceImpl(memory);
+        const observableSource = new ObservableSource(memory);
         const fork = new PassThroughFork(observableSource);
         const authentication = new MockAuthentication(memory);
         const factManager = new FactManager(authentication, fork, observableSource);
