+++ test/watch/watchSpec.ts
@@ -3,7 +3,7 @@ import { PassThroughFork } from "../../src/fork/pass-through-fork";
 import { Jinaga } from "../../src/jinaga";
 import { FactManager } from "../../src/managers/factManager";
 import { MemoryStore } from "../../src/memory/memory-store";
-import { ObservableSourceImpl } from "../../src/observable/observable-source-impl";
+import { ObservableSource } from "../../src/observable/observable";
 import { factReferenceEquals } from "../../src/storage";
 import { MockAuthentication } from "./mock-authentication";
 
@@ -56,7 +56,7 @@ describe("Watch", () => {
   var j: Jinaga;
   beforeEach(() => {
     const memory = new MemoryStore();
-    const observableSource = new ObservableSourceImpl(memory);
+    const observableSource = new ObservableSource(memory);
     const fork = new PassThroughFork(observableSource);
     const authentication = new MockAuthentication(memory);
     const factManager = new FactManager(authentication, fork, observableSource);
